requirements.txt
Django==4.1.5
django-crispy-forms==1.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
transliterate==1.10.2
Pillow==9.4.0
